feat: add session menu back
